add: defensive programming. CUD tasks only if superuser

+++ products/views.py
@@ -80,9 +80,12 @@ def product_detail(request, product_id):
 
 @login_required
 def add_product(request):
-    """ Add a product to the store """
+    """
+    Add a product to the store
+    """
     if not request.user.is_superuser:
-        messages.error(request, 'Sorry, only store owners can do that.')
+        messages.error(request, 'Sorry, only Administrators \
+            can perform that task.')
         return redirect(reverse('home'))
 
     if request.method == 'POST':
@@ -110,6 +113,11 @@ def edit_product(request, product_id):
     """
     Edit a products details
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Administrators \
+            can perform that task.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -139,6 +147,11 @@ def delete_product(request, product_id):
     """
     Delete Product
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Administrators can \
+            perform that task.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'Product {product.name} \
